+++ src/Kiota.Builder/Configuration/GenerationConfiguration.cs
@@ -40,6 +40,10 @@ public class GenerationConfiguration : ICloneable
     public string ClientClassName { get; set; } = "ApiClient";
     public string ClientNamespaceName { get; set; } = "ApiSdk";
     public string NamespaceNameSeparator { get; set; } = ".";
+    public bool ExportPublicApi
+    {
+        get; set;
+    }
     internal const string ModelsNamespaceSegmentName = "models";
     public string ModelsNamespaceName
     {
@@ -153,6 +157,7 @@ public class GenerationConfiguration : ICloneable
             PatternsOverride = new(PatternsOverride ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
             PluginTypes = new(PluginTypes ?? Enumerable.Empty<PluginType>()),
             DisableSSLValidation = DisableSSLValidation,
+            ExportPublicApi = ExportPublicApi,
             PluginAuthInformation = PluginAuthInformation,
         };
     }