+++ tests/Kiota.Builder.Tests/Plugins/PluginsGenerationServiceTests.cs
@@ -15,7 +15,6 @@ using Moq;
 using Xunit;
 
 namespace Kiota.Builder.Tests.Plugins;
-
 public sealed class PluginsGenerationServiceTests : IDisposable
 {
     private readonly HttpClient _httpClient = new();
@@ -23,13 +22,10 @@ public sealed class PluginsGenerationServiceTests : IDisposable
     [Fact]
     public void Defensive()
     {
-        Assert.Throws<ArgumentNullException>(() =>
-            new PluginsGenerationService(null, OpenApiUrlTreeNode.Create(), new(), "foo"));
+        Assert.Throws<ArgumentNullException>(() => new PluginsGenerationService(null, OpenApiUrlTreeNode.Create(), new(), "foo"));
         Assert.Throws<ArgumentNullException>(() => new PluginsGenerationService(new(), null, new(), "foo"));
-        Assert.Throws<ArgumentNullException>(() =>
-            new PluginsGenerationService(new(), OpenApiUrlTreeNode.Create(), null, "foo"));
-        Assert.Throws<ArgumentException>(() =>
-            new PluginsGenerationService(new(), OpenApiUrlTreeNode.Create(), new(), string.Empty));
+        Assert.Throws<ArgumentNullException>(() => new PluginsGenerationService(new(), OpenApiUrlTreeNode.Create(), null, "foo"));
+        Assert.Throws<ArgumentException>(() => new PluginsGenerationService(new(), OpenApiUrlTreeNode.Create(), new(), string.Empty));
     }
 
     public void Dispose()
@@ -39,8 +35,8 @@ public sealed class PluginsGenerationServiceTests : IDisposable
 
     [Theory]
     [InlineData("client", "client")]
-    [InlineData("Budget Tracker", "BudgetTracker")] //drop the space
-    [InlineData("My-Super complex() %@#$& Name", "MySupercomplexName")] //drop the space and special characters
+    [InlineData("Budget Tracker", "BudgetTracker")]//drop the space
+    [InlineData("My-Super complex() %@#$& Name", "MySupercomplexName")]//drop the space and special characters
     public async Task GeneratesManifest(string inputPluginName, string expectedPluginName)
     {
         var simpleDescriptionContent = @"openapi: 3.0.0
@@ -86,15 +82,12 @@ paths:
             ClientClassName = inputPluginName,
             ApiRootUrl = "http://localhost/", //Kiota builder would set this for us
         };
-        var (openAPIDocumentStream, _) =
-            await openAPIDocumentDS.LoadStreamAsync(simpleDescriptionPath, generationConfiguration, null, false);
-        var openApiDocument =
-            await openAPIDocumentDS.GetDocumentFromStreamAsync(openAPIDocumentStream, generationConfiguration);
+        var (openAPIDocumentStream, _) = await openAPIDocumentDS.LoadStreamAsync(simpleDescriptionPath, generationConfiguration, null, false);
+        var openApiDocument = await openAPIDocumentDS.GetDocumentFromStreamAsync(openAPIDocumentStream, generationConfiguration);
         KiotaBuilder.CleanupOperationIdForPlugins(openApiDocument);
         var urlTreeNode = OpenApiUrlTreeNode.Create(openApiDocument, Constants.DefaultOpenApiLabel);
 
-        var pluginsGenerationService =
-            new PluginsGenerationService(openApiDocument, urlTreeNode, generationConfiguration, workingDirectory);
+        var pluginsGenerationService = new PluginsGenerationService(openApiDocument, urlTreeNode, generationConfiguration, workingDirectory);
         await pluginsGenerationService.GenerateManifestAsync();
 
         Assert.True(File.Exists(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-apiplugin.json")));
@@ -107,26 +100,21 @@ paths:
         Assert.False(File.Exists(Path.Combine(outputDirectory, "outline.png")));
 
         // Validate the v2 plugin
-        var manifestContent =
-            await File.ReadAllTextAsync(Path.Combine(outputDirectory,
-                $"{expectedPluginName.ToLower()}-apiplugin.json"));
+        var manifestContent = await File.ReadAllTextAsync(Path.Combine(outputDirectory, $"{expectedPluginName.ToLower()}-apiplugin.json"));
         using var jsonDocument = JsonDocument.Parse(manifestContent);
         var resultingManifest = PluginManifestDocument.Load(jsonDocument.RootElement);
         Assert.NotNull(resultingManifest.Document);
-        Assert.Equal($"{expectedPluginName.ToLower()}-openapi.yml",
-            resultingManifest.Document.Runtimes.OfType<OpenApiRuntime>().First().Spec.Url);
-        Assert.Equal(2,
-            resultingManifest.Document.Functions.Count); // all functions are generated despite missing operationIds
-        Assert.Equal(expectedPluginName, resultingManifest.Document.Namespace); // namespace is cleaned up.
-        Assert.Empty(resultingManifest.Problems); // no problems are expected with names
+        Assert.Equal($"{expectedPluginName.ToLower()}-openapi.yml", resultingManifest.Document.Runtimes.OfType<OpenApiRuntime>().First().Spec.Url);
+        Assert.Equal(2, resultingManifest.Document.Functions.Count);// all functions are generated despite missing operationIds
+        Assert.Equal(expectedPluginName, resultingManifest.Document.Namespace);// namespace is cleaned up.
+        Assert.Empty(resultingManifest.Problems);// no problems are expected with names
     }
-
     private const string ManifestFileName = "client-apiplugin.json";
     private const string OpenAIPluginFileName = "openai-plugins.json";
     private const string OpenApiFileName = "client-openapi.yml";
 
     [Fact]
-    public async Task GeneratesManifestAndCleansUpInputDescription()
+    public async Task GeneratesManifestAndCleansUpInputDescriptionAsync()
     {
         var simpleDescriptionContent = @"openapi: 3.0.0
 info:
@@ -190,15 +178,12 @@ components:
             ClientClassName = "client",
             ApiRootUrl = "http://localhost/", //Kiota builder would set this for us
         };
-        var (openAPIDocumentStream, _) =
-            await openAPIDocumentDS.LoadStreamAsync(simpleDescriptionPath, generationConfiguration, null, false);
-        var openApiDocument =
-            await openAPIDocumentDS.GetDocumentFromStreamAsync(openAPIDocumentStream, generationConfiguration);
+        var (openAPIDocumentStream, _) = await openAPIDocumentDS.LoadStreamAsync(simpleDescriptionPath, generationConfiguration, null, false);
+        var openApiDocument = await openAPIDocumentDS.GetDocumentFromStreamAsync(openAPIDocumentStream, generationConfiguration);
         KiotaBuilder.CleanupOperationIdForPlugins(openApiDocument);
         var urlTreeNode = OpenApiUrlTreeNode.Create(openApiDocument, Constants.DefaultOpenApiLabel);
 
-        var pluginsGenerationService =
-            new PluginsGenerationService(openApiDocument, urlTreeNode, generationConfiguration, workingDirectory);
+        var pluginsGenerationService = new PluginsGenerationService(openApiDocument, urlTreeNode, generationConfiguration, workingDirectory);
         await pluginsGenerationService.GenerateManifestAsync();
 
         Assert.True(File.Exists(Path.Combine(outputDirectory, ManifestFileName)));
@@ -210,9 +195,8 @@ components:
         var resultingManifest = PluginManifestDocument.Load(jsonDocument.RootElement);
         Assert.NotNull(resultingManifest.Document);
         Assert.Equal(OpenApiFileName, resultingManifest.Document.Runtimes.OfType<OpenApiRuntime>().First().Spec.Url);
-        Assert.Equal(2,
-            resultingManifest.Document.Functions.Count); // all functions are generated despite missing operationIds
-        Assert.Empty(resultingManifest.Problems); // no problems are expected with names
+        Assert.Equal(2, resultingManifest.Document.Functions.Count);// all functions are generated despite missing operationIds
+        Assert.Empty(resultingManifest.Problems);// no problems are expected with names
 
         var openApiReader = new OpenApiStreamReader();
 
@@ -221,22 +205,15 @@ components:
         var originalDocument = openApiReader.Read(originalOpenApiFile, out var originalDiagnostic);
         Assert.Empty(originalDiagnostic.Errors);
 
-        Assert.Equal(originalDocument.Paths["/test"].Operations[OperationType.Get].Description,
-            resultingManifest.Document.Functions[0].Description); // pulls from description
-        Assert.Equal(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Summary,
-            resultingManifest.Document.Functions[1].Description); // pulls from summary
-        Assert.Single(originalDocument.Components.Schemas); // one schema originally
+        Assert.Equal(originalDocument.Paths["/test"].Operations[OperationType.Get].Description, resultingManifest.Document.Functions[0].Description);// pulls from description
+        Assert.Equal(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Summary, resultingManifest.Document.Functions[1].Description);// pulls from summary
+        Assert.Single(originalDocument.Components.Schemas);// one schema originally
         Assert.Single(originalDocument.Extensions); // single unsupported extension at root
         Assert.Equal(2, originalDocument.Paths.Count); // document has only two paths
-        Assert.Equal(2,
-            originalDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count); // 2 responses originally
-        Assert.Single(originalDocument.Paths["/test"].Operations[OperationType.Get]
-            .Extensions); // 1 UNsupported extension
-        Assert.Equal(2,
-            originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses
-                .Count); // 2 responses originally
-        Assert.Single(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get]
-            .Extensions); // 1 supported extension
+        Assert.Equal(2, originalDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count); // 2 responses originally
+        Assert.Single(originalDocument.Paths["/test"].Operations[OperationType.Get].Extensions); // 1 UNsupported extension
+        Assert.Equal(2, originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses.Count); // 2 responses originally
+        Assert.Single(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Extensions); // 1 supported extension
 
         // Validate the output open api file
         var resultOpenApiFile = File.OpenRead(Path.Combine(outputDirectory, OpenApiFileName));
@@ -244,24 +221,17 @@ components:
         Assert.Empty(diagnostic.Errors);
 
         // Assertions / validations
-        Assert.Empty(resultDocument.Components.Schemas); // no schema is referenced. so ensure they are all removed
+        Assert.Empty(resultDocument.Components.Schemas);// no schema is referenced. so ensure they are all removed
         Assert.Empty(resultDocument.Extensions); // no extension at root (unsupported extension is removed)
         Assert.Equal(2, resultDocument.Paths.Count); // document has only two paths
-        Assert.Equal(originalDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count,
-            resultDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count); // Responses are still intact.
-        Assert.NotEmpty(resultDocument.Paths["/test"].Operations[OperationType.Get].Responses["200"]
-            .Description); // response description string is not empty
-        Assert.Empty(resultDocument.Paths["/test"].Operations[OperationType.Get]
-            .Extensions); // NO UNsupported extension
-        Assert.Equal(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses.Count,
-            resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses
-                .Count); // Responses are still intact.
-        Assert.NotEmpty(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses["200"]
-            .Description); // response description string is not empty
-        Assert.Single(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get]
-            .Extensions); // 1 supported extension still present in operation
+        Assert.Equal(originalDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count, resultDocument.Paths["/test"].Operations[OperationType.Get].Responses.Count); // Responses are still intact.
+        Assert.NotEmpty(resultDocument.Paths["/test"].Operations[OperationType.Get].Responses["200"].Description); // response description string is not empty
+        Assert.Empty(resultDocument.Paths["/test"].Operations[OperationType.Get].Extensions); // NO UNsupported extension
+        Assert.Equal(originalDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses.Count, resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses.Count); // Responses are still intact.
+        Assert.NotEmpty(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Responses["200"].Description);// response description string is not empty
+        Assert.Single(resultDocument.Paths["/test/{id}"].Operations[OperationType.Get].Extensions); // 1 supported extension still present in operation
     }
-
+    
     public static TheoryData<string, string, string, PluginAuthConfiguration, Action<DocumentValidationResults<PluginManifestDocument>>>
         SecurityInformationSuccess()
     {